// src/types/crypto.ts
export type SortField = 'market_cap_rank' | 'current_price' | 'price_change_percentage_24h' | 'market_cap' | 'total_volume';
export type SortOrder = 'asc' | 'desc';